refactor(builder): break pipeline cycle by extracting orchestrator-selection strategy
Extract the native-orchestrator path out of pipeline.ts into two new stage
modules:

  - stages/native-orchestrator.ts — tryNativeOrchestrator + post-native
    structure/analysis fallback + dropped-language detection/backfill.
  - stages/native-db-lifecycle.ts — shared rusqlite connection helpers
    (closeNativeDb, reopenNativeDb, suspendNativeDb, refreshJsDb).

This breaks the function-level cycle 'buildGraph <-> tryNativeOrchestrator'
caused by codegraph's name-based resolver conflating the local buildGraph
function with the ctx.nativeDb.buildGraph() method call. Once the
orchestrator lives in its own file, there is no longer a local buildGraph
in scope to collide with the method invocation.

Function-level cycles: 9 -> 5. No file-level cycle introduced (still 1,
unchanged — pre-existing MCP cycle). pipeline.ts shrinks from 1404 to 465
lines and now reads as a thin top-level controller: detect changes, try
native, fall back to JS stages.

computeWasmOnlyStaleFiles is re-exported from pipeline.ts so existing
unit tests (tests/builder/wasm-only-stale-files.test.ts) keep working
without changes.
